// src/main/resources/angular/src/app/api/model/vehicleDGTRes.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: kfiertek@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface VehicleDGTRes { 
    /**
     * 0,EC,B,C
     */
    ecoTag: string;
    codiIndustria: string;
    tipoVehiculoDGT: string;
    tipoVehiculoDUM: string;
    marca: string;
    modelo: string;
    cif: string;
    dni: string;
}